refactor(generator): move AssemblyConventions to Discovery namespace
